+++ apps/demo-app/src/components/layout-demo/MobileBottomFlyInDemo.vue
@@ -33,7 +33,7 @@ defineSlots<{
   max-height: fit-content;
   margin: auto;
 
-  border: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
+  border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
   background-color: var(--onyx-color-base-background-blank);
 }
 </style>