+++ apps/demo-app/src/components/layout-demo/SidebarDemo.vue
@@ -38,7 +38,7 @@ defineSlots<{
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  border-right: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
+  border-right: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
 
   &__close {
     justify-self: flex-end;