+++ packages/sit-onyx/src/components/OnyxAvatar/OnyxAvatar.ct.tsx
@@ -25,8 +25,6 @@ test.describe("Screenshot tests", () => {
     name: "Avatar",
     columns: ["default", "custom"],
     rows: AVATAR_SIZES,
-    // TODO: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
-    disabledAccessibilityRules: ["color-contrast"],
     component: (column, row) => (
       <OnyxAvatar
         label="John Doe"
@@ -43,8 +41,6 @@ test.describe("Screenshot tests", () => {
     name: "Avatar (custom content)",
     columns: ["default", "truncation"],
     rows: AVATAR_SIZES,
-    // TODO: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
-    disabledAccessibilityRules: ["color-contrast"],
     component: (column, row) => (
       <OnyxAvatar label="Custom content" size={row}>
         {column === "truncation" ? "+999999" : "+42"}