+++ packages/sit-onyx/src/components/OnyxCheckboxGroup/OnyxCheckboxGroup.vue
@@ -119,7 +119,7 @@ const checkAllLabel = computed(() => {
     border: none;
     max-width: max-content;
     min-width: unset;
-    $check-all-border: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
+    $check-all-border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
 
     &__label {
       display: flex;