+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.vue
@@ -1,22 +1,27 @@
 <script lang="ts" setup>
+import checkSmall from "@sit-onyx/icons/check-small.svg?raw";
+import xSmall from "@sit-onyx/icons/x-small.svg?raw";
 import { computed } from "vue";
 import { useDensity } from "../../composables/density";
 import { getFormMessages, useCustomValidity } from "../../composables/useCustomValidity";
 import { useErrorClass } from "../../composables/useErrorClass";
 import { SKELETON_INJECTED_SYMBOL, useSkeletonContext } from "../../composables/useSkeletonState";
+import { injectI18n } from "../../i18n";
 import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm.core";
 import OnyxFormElement from "../OnyxFormElement/OnyxFormElement.vue";
+import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
 import OnyxLoadingIndicator from "../OnyxLoadingIndicator/OnyxLoadingIndicator.vue";
 import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
 import type { OnyxInputProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxInputProps>(), {
-  modelValue: "",
   type: "text",
   required: false,
   autocapitalize: "sentences",
   readonly: false,
   loading: false,
+  hideClearIcon: false,
+  hideSuccessIcon: false,
   skeleton: SKELETON_INJECTED_SYMBOL,
   disabled: FORM_INJECTED_SYMBOL,
   showError: FORM_INJECTED_SYMBOL,
@@ -33,6 +38,7 @@ const emit = defineEmits<{
   validityChange: [validity: ValidityState];
 }>();
 
+const { t } = injectI18n();
 const { vCustomValidity, errorMessages } = useCustomValidity({ props, emit });
 const successMessages = computed(() => getFormMessages(props.success));
 const messages = computed(() => getFormMessages(props.message));
@@ -42,10 +48,7 @@ const { densityClass } = useDensity(props);
 /**
  * Current value (with getter and setter) that can be used as "v-model" for the native input.
  */
-const value = computed({
-  get: () => props.modelValue,
-  set: (value) => emit("update:modelValue", value),
-});
+const value = defineModel<string>({ default: "" });
 
 const patternSource = computed(() => {
   if (props.pattern instanceof RegExp) return props.pattern.source;
@@ -93,6 +96,26 @@ const errorClass = useErrorClass(showError);
             :aria-label="props.hideLabel ? props.label : undefined"
             :title="props.hideLabel ? props.label : undefined"
           />
+
+          <button
+            v-if="!props.hideClearIcon && value !== ''"
+            type="button"
+            class="onyx-input__clear"
+            :aria-label="t('input.clear')"
+            :title="t('input.clear')"
+            @click="() => (value = '')"
+          >
+            <OnyxIcon :icon="xSmall" />
+          </button>
+
+          <OnyxIcon
+            v-if="!props.hideSuccessIcon && successMessages"
+            class="onyx-input__check-icon"
+            :icon="checkSmall"
+            color="success"
+          />
+
+          <!-- eslint-enable vuejs-accessibility/no-autofocus -->
         </div>
       </template>
     </OnyxFormElement>
@@ -120,6 +143,41 @@ const errorClass = useErrorClass(showError);
       $base-selector: ".onyx-input",
       $vertical-padding: var(--onyx-input-padding-vertical)
     );
+  }
+
+  ::-webkit-search-cancel-button {
+    display: none;
+  }
+
+  &__clear {
+    height: 100%;
+    margin: 0;
+    padding: 0;
+
+    display: flex;
+    align-items: center;
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+    color: var(--onyx-color-text-icons-neutral-intense);
+
+    &:hover,
+    &:focus-visible {
+      color: var(--onyx-color-text-icons-primary-intense);
+    }
+
+    &:focus-visible {
+      outline: var(--onyx-outline-width) solid var(--onyx-color-component-focus-primary);
+      border-radius: var(--onyx-radius-sm);
+    }
+  }
+
+  // hide clear icon when input is not focussed
+  &:not(&:has(&__wrapper:focus-within)),
+  &:has(&__native:read-only) {
+    .onyx-input__clear {
+      display: none;
+    }
   }
 }
 </style>