+++ packages/sit-onyx/src/components/OnyxNavBar/modules/OnyxNavButton/OnyxNavButton.vue
@@ -131,7 +131,7 @@ $border-radius: var(--onyx-radius-sm);
       border: none;
 
       &:focus-visible {
-        outline: var(--onyx-outline-width) solid var(--onyx-color-base-primary-200);
+        outline: var(--onyx-outline-width) solid var(--onyx-color-component-focus-primary);
       }
     }
 
@@ -195,11 +195,15 @@ $border-radius: var(--onyx-radius-sm);
         padding-inline: $mobile-padding;
       }
 
+      &:not(.onyx-nav-button--active) .onyx-nav-button__trigger {
+        color: var(--onyx-color-text-icons-neutral-intense);
+      }
+
       > .onyx-nav-button__trigger,
       .onyx-menu-item {
         width: 100%;
         justify-content: flex-start;
-        border: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
+        border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
         height: 3rem;
         border-radius: $border-radius;
 
@@ -213,7 +217,6 @@ $border-radius: var(--onyx-radius-sm);
 
         &:hover {
           background-color: var(--onyx-color-base-background-tinted);
-          border-color: var(--onyx-color-base-primary-500);
         }
       }
 
@@ -231,7 +234,7 @@ $border-radius: var(--onyx-radius-sm);
         font-weight: 600;
 
         &:hover {
-          border-color: var(--onyx-color-base-primary-500);
+          border-color: var(--onyx-color-component-border-primary);
           background-color: var(--onyx-list-item-background-selected);
         }
       }