+++ packages/sit-onyx/src/components/OnyxSelect/OnyxSelect.vue
@@ -508,11 +508,11 @@ const selectInputProps = computed(() => {
     }
 
     &__search {
-      border-bottom: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
+      border-bottom: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
     }
 
     &:has(&__wrapper:focus-visible) {
-      outline: var(--onyx-outline-width) solid var(--onyx-color-base-primary-200);
+      outline: var(--onyx-outline-width) solid var(--onyx-color-component-focus-primary);
     }
 
     &__wrapper:has(.onyx-mini-search) {