+++ packages/sit-onyx/src/components/examples/GridPlayground/GridBadge/GridBadge.ct.tsx
@@ -7,8 +7,6 @@ test.describe("screenshot tests", () => {
     name: "Grid badge",
     columns: ["default"],
     rows: ["info", "warning", "danger"],
-    // TODO: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
-    disabledAccessibilityRules: ["color-contrast"],
     component: (column, row) => <GridBadge label="Label" value="value" color={row} />,
     beforeScreenshot: async (component) => {
       await expect(component.getByLabel("Label")).toBeAttached();