+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,40 @@
 # playground
 
+## 1.0.0-beta.86
+
+### Patch Changes
+
+- Updated dependencies [c358978]
+  - sit-onyx@1.0.0-beta.86
+
+## 1.0.0-beta.85
+
+### Patch Changes
+
+- Updated dependencies [1a11a27]
+  - sit-onyx@1.0.0-beta.85
+
+## 1.0.0-beta.84
+
+### Patch Changes
+
+- Updated dependencies [2a7a285]
+  - sit-onyx@1.0.0-beta.84
+
+## 1.0.0-beta.83
+
+### Patch Changes
+
+- Updated dependencies [8fe041f]
+  - sit-onyx@1.0.0-beta.83
+
+## 1.0.0-beta.82
+
+### Patch Changes
+
+- Updated dependencies [5fba96d]
+  - sit-onyx@1.0.0-beta.82
+
 ## 1.0.0-beta.81
 
 ### Patch Changes